backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langgraph
pydantic-ai
google-generativeai
presidio-analyzer
presidio-anonymizer